testing/20150817.html

<!DOCTYPE html>
<title>IRC logs: w3c / #testing / 20150817</title>
<meta name="viewport" content="width=device-width">
<style>
ol { font-family: monospace; color: #000; font-size: 90%; list-style: none; margin: 0; padding: 0; }
li { margin: 0; padding: 0; white-space: pre-wrap; word-wrap: break-word; }
li:target { background: #f90; }
li.hide { display: none; }
li.flagged { background: yellow; }
li.filtered { color: #aaa; font-size: 60%; }
li.filtered:hover { font-size: 100%; color: #000; height: auto; }
li span { display: inline; margin: 0; padding: 0 10px; background: #eee; color: #000; text-decoration: none; cursor: pointer; }
li.flagged span { background: #f00; }
li:hover span { background: #ff0 }
input#filter { margin: 0 10px; }
label { cursor: pointer; }
p span { background: #ff0; margin-left: 10px; }
</style>
<h1 id="w3c/testing.20150817.log"><a href="/irc-logs/">/irc-logs</a> / w3c / <a href="irc://irc.w3.org:6665/testing">#testing</a> / 2015-08-17 / <a href="#end">end</a></h1>
<p>Options: 
<p><a href=20150816>Previous day</a>, <a href=20150818>Next day</a></p>
<ol id=lines>
<li id=l-1 class=hide><a href=#l-1>#</a> Session Start: Mon Aug 17 00:00:00 2015</li>
<li id=l-2 class=hide><a href=#l-2>#</a> Session Ident: #testing</li>
<li id=l-3 class=hide><a href=#l-3>#</a> [00:10] * Quits: jyasskin (~textual@public.cloak) (&quot;My computer has gone to sleep. ZZZzzz…&quot;)</li>
<li id=l-4 class=hide><a href=#l-4>#</a> [01:26] * Joins: jyasskin (~textual@public.cloak)</li>
<li id=l-5 class=hide><a href=#l-5>#</a> [02:19] * Quits: jyasskin (~textual@public.cloak) (&quot;My computer has gone to sleep. ZZZzzz…&quot;)</li>
<li id=l-6 class=hide><a href=#l-6>#</a> [02:22] * Joins: bhill2 (~bhill2@public.cloak)</li>
<li id=l-7 class=hide><a href=#l-7>#</a> [02:25] * Quits: bhill2 (~bhill2@public.cloak) (Client closed connection)</li>
<li id=l-8 class=hide><a href=#l-8>#</a> [02:26] * Joins: bhill2 (~bhill2@public.cloak)</li>
<li id=l-9 class=hide><a href=#l-9>#</a> [02:39] * Quits: bhill2 (~bhill2@public.cloak) (Ping timeout: 180 seconds)</li>
<li id=l-10 class=hide><a href=#l-10>#</a> [03:15] * Joins: jyasskin (~textual@public.cloak)</li>
<li id=l-11 class=hide><a href=#l-11>#</a> [03:29] * Quits: jyasskin (~textual@public.cloak) (&quot;My computer has gone to sleep. ZZZzzz…&quot;)</li>
<li id=l-12 class=hide><a href=#l-12>#</a> [03:51] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-13><a href=#l-13>#</a> [03:51] -gitbot:#testing- [testharness.js] sideshowbarker created sideshowbarker/assert_object_equals (+1 new commit): <a href="https://github.com/w3c/testharness.js/commit/0214d0b19597">https://github.com/w3c/testharness.js/commit/0214d0b19597</a></li>
<li id=l-14><a href=#l-14>#</a> [03:51] -gitbot:#testing- testharness.js/sideshowbarker/assert_object_equals 0214d0b Michael[tm] Smith: Remove assert_object_equals....</li>
<li id=l-15 class=hide><a href=#l-15>#</a> [03:51] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-16 class=hide><a href=#l-16>#</a> [03:52] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-17><a href=#l-17>#</a> [03:52] -gitbot:#testing- [testharness.js] sideshowbarker opened pull request #143: Remove assert_object_equals. (master...sideshowbarker/assert_object_equals) <a href="https://github.com/w3c/testharness.js/pull/143">https://github.com/w3c/testharness.js/pull/143</a></li>
<li id=l-18 class=hide><a href=#l-18>#</a> [03:52] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-19 class=hide><a href=#l-19>#</a> [04:10] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-20><a href=#l-20>#</a> [04:10] -gitbot:#testing- [web-platform-tests] sideshowbarker created sideshowbarker/assert_object_equals (+1 new commit): <a href="https://github.com/w3c/web-platform-tests/commit/fcda8fc274c8">https://github.com/w3c/web-platform-tests/commit/fcda8fc274c8</a></li>
<li id=l-21><a href=#l-21>#</a> [04:10] -gitbot:#testing- web-platform-tests/sideshowbarker/assert_object_equals fcda8fc Michael[tm] Smith: Replaced a simple assert_object_equals case....</li>
<li id=l-22 class=hide><a href=#l-22>#</a> [04:10] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-23 class=hide><a href=#l-23>#</a> [04:10] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-24><a href=#l-24>#</a> [04:10] -gitbot:#testing- [web-platform-tests] sideshowbarker opened pull request #2086: Replaced a simple assert_object_equals case. (master...sideshowbarker/assert_object_equals) <a href="https://github.com/w3c/web-platform-tests/pull/2086">https://github.com/w3c/web-platform-tests/pull/2086</a></li>
<li id=l-25 class=hide><a href=#l-25>#</a> [04:10] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-26 class=hide _a="heycam|away"><a href=#l-26>#</a> [04:57] * heycam|away is now known as heycam</li>
<li id=l-27 class=hide><a href=#l-27>#</a> [05:09] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-28><a href=#l-28>#</a> [05:09] -gitbot:#testing- [web-platform-tests] sideshowbarker closed pull request #2086: Replaced a simple assert_object_equals case. (master...sideshowbarker/assert_object_equals) <a href="https://github.com/w3c/web-platform-tests/pull/2086">https://github.com/w3c/web-platform-tests/pull/2086</a></li>
<li id=l-29 class=hide><a href=#l-29>#</a> [05:09] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-30 class=hide><a href=#l-30>#</a> [05:09] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-31><a href=#l-31>#</a> [05:09] -gitbot:#testing- [web-platform-tests] sideshowbarker deleted sideshowbarker/assert_object_equals at fcda8fc: <a href="https://github.com/w3c/web-platform-tests/commit/fcda8fc">https://github.com/w3c/web-platform-tests/commit/fcda8fc</a></li>
<li id=l-32 class=hide><a href=#l-32>#</a> [05:09] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-33 class=hide><a href=#l-33>#</a> [05:09] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-34><a href=#l-34>#</a> [05:09] -gitbot:#testing- [web-platform-tests] sideshowbarker created sideshowbarker/assert_object_equals (+1 new commit): <a href="https://github.com/w3c/web-platform-tests/commit/1cc2533cd321">https://github.com/w3c/web-platform-tests/commit/1cc2533cd321</a></li>
<li id=l-35><a href=#l-35>#</a> [05:09] -gitbot:#testing- web-platform-tests/sideshowbarker/assert_object_equals 1cc2533 Michael[tm] Smith: Replaced some simple assert_object_equals cases....</li>
<li id=l-36 class=hide><a href=#l-36>#</a> [05:09] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-37 class=hide><a href=#l-37>#</a> [05:10] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-38><a href=#l-38>#</a> [05:10] -gitbot:#testing- [web-platform-tests] sideshowbarker opened pull request #2088: Replaced some simple assert_object_equals cases. (master...sideshowbarker/assert_object_equals) <a href="https://github.com/w3c/web-platform-tests/pull/2088">https://github.com/w3c/web-platform-tests/pull/2088</a></li>
<li id=l-39 class=hide><a href=#l-39>#</a> [05:10] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-40 class=hide><a href=#l-40>#</a> [05:46] * Joins: gamekathu (~gamekathu@public.cloak)</li>
<li id=l-41 class=hide><a href=#l-41>#</a> [05:52] * Joins: MK (kawamori@public.cloak)</li>
<li id=l-42 class=hide><a href=#l-42>#</a> [06:12] * Joins: cvrebert (~cvrebert@public.cloak)</li>
<li id=l-43 class=hide><a href=#l-43>#</a> [06:12] * Quits: kochi (~kochi@public.cloak) (&quot;Leaving.&quot;)</li>
<li id=l-44 _a="cvrebert" _r="botie"><a href=#l-44>#</a> [06:13] &lt;cvrebert&gt; botie: any news?</li>
<li id=l-45 _a="botie" _r="cvrebert"><a href=#l-45>#</a> [06:13] &lt;botie&gt; cvrebert: no idea</li>
<li id=l-46 class=hide><a href=#l-46>#</a> [06:13] * Quits: cvrebert (~cvrebert@public.cloak) (cvrebert)</li>
<li id=l-47 class=hide><a href=#l-47>#</a> [06:15] * Joins: kochi (~kochi@public.cloak)</li>
<li id=l-48 class=hide><a href=#l-48>#</a> [06:32] * Quits: MK (kawamori@public.cloak) (Client closed connection)</li>
<li id=l-49 class=hide><a href=#l-49>#</a> [06:52] * Joins: kochi1 (~kochi@public.cloak)</li>
<li id=l-50 class=hide><a href=#l-50>#</a> [06:52] * Quits: kochi1 (~kochi@public.cloak) (&quot;Leaving.&quot;)</li>
<li id=l-51 class=hide><a href=#l-51>#</a> [06:52] * Joins: kochi1 (~kochi@public.cloak)</li>
<li id=l-52 class=hide><a href=#l-52>#</a> [06:53] * Quits: kochi1 (~kochi@public.cloak) (&quot;Leaving.&quot;)</li>
<li id=l-53 class=hide><a href=#l-53>#</a> [06:53] * Joins: kochi1 (~kochi@public.cloak)</li>
<li id=l-54 class=hide><a href=#l-54>#</a> [06:53] * Quits: kochi1 (~kochi@public.cloak) (&quot;Leaving.&quot;)</li>
<li id=l-55 class=hide><a href=#l-55>#</a> [06:53] * Joins: kochi1 (~kochi@public.cloak)</li>
<li id=l-56 class=hide _a="heycam"><a href=#l-56>#</a> [07:15] * heycam is now known as heycam|away</li>
<li id=l-57 class=hide><a href=#l-57>#</a> [07:19] * Quits: kochi (~kochi@public.cloak) (&quot;Leaving.&quot;)</li>
<li id=l-58 class=hide><a href=#l-58>#</a> [07:20] * Quits: kochi1 (~kochi@public.cloak) (Client closed connection)</li>
<li id=l-59 class=hide><a href=#l-59>#</a> [07:22] * Joins: kochi (~kochi@public.cloak)</li>
<li id=l-60 class=hide><a href=#l-60>#</a> [07:32] * Joins: kochi1 (~kochi@public.cloak)</li>
<li id=l-61 class=hide><a href=#l-61>#</a> [07:32] * Quits: kochi (~kochi@public.cloak) (Client closed connection)</li>
<li id=l-62 class=hide><a href=#l-62>#</a> [07:34] * Quits: kochi1 (~kochi@public.cloak) (Client closed connection)</li>
<li id=l-63 class=hide><a href=#l-63>#</a> [07:46] * Joins: MK (kawamori@public.cloak)</li>
<li id=l-64 class=hide _a="heycam|away"><a href=#l-64>#</a> [07:51] * heycam|away is now known as heycam</li>
<li id=l-65 class=hide><a href=#l-65>#</a> [08:02] * Joins: gool (~gool@public.cloak)</li>
<li id=l-66 class=hide><a href=#l-66>#</a> [08:19] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-67><a href=#l-67>#</a> [08:19] -gitbot:#testing- [testharness.js] inexorabletash closed pull request #139: assert_object_equals: Check property ownership (master...inexorabletash/object-equals-own) <a href="https://github.com/w3c/testharness.js/pull/139">https://github.com/w3c/testharness.js/pull/139</a></li>
<li id=l-68 class=hide><a href=#l-68>#</a> [08:19] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-69 class=hide><a href=#l-69>#</a> [08:34] * Quits: MK (kawamori@public.cloak) (Ping timeout: 180 seconds)</li>
<li id=l-70 class=hide><a href=#l-70>#</a> [08:35] * Joins: MK (kawamori@public.cloak)</li>
<li id=l-71 class=hide><a href=#l-71>#</a> [08:39] * Quits: gool (~gool@public.cloak) (&quot;Page closed&quot;)</li>
<li id=l-72 class=hide><a href=#l-72>#</a> [08:43] * Joins: jyasskin (~textual@public.cloak)</li>
<li id=l-73 class=hide><a href=#l-73>#</a> [08:49] * Quits: jyasskin (~textual@public.cloak) (&quot;My computer has gone to sleep. ZZZzzz…&quot;)</li>
<li id=l-74 class=hide><a href=#l-74>#</a> [09:25] * Quits: MK (kawamori@public.cloak) (Ping timeout: 180 seconds)</li>
<li id=l-75 class=hide><a href=#l-75>#</a> [09:36] * Joins: MK (kawamori@public.cloak)</li>
<li id=l-76 class=hide><a href=#l-76>#</a> [09:36] * Joins: Ms2ger (~Ms2ger@public.cloak)</li>
<li id=l-77 class=hide><a href=#l-77>#</a> [10:06] * Joins: dom (dom@public.cloak)</li>
<li id=l-78 class=hide><a href=#l-78>#</a> [10:24] * Quits: MK (kawamori@public.cloak) (Ping timeout: 180 seconds)</li>
<li id=l-79 class=hide _a="heycam"><a href=#l-79>#</a> [10:26] * heycam is now known as heycam|away</li>
<li id=l-80 class=hide><a href=#l-80>#</a> [10:32] * Joins: Lachy (~Lachy@public.cloak)</li>
<li id=l-81 class=hide><a href=#l-81>#</a> [10:33] * Joins: MK (kawamori@public.cloak)</li>
<li id=l-82 class=hide><a href=#l-82>#</a> [10:35] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-83><a href=#l-83>#</a> [10:35] -gitbot:#testing- [web-platform-tests] TakayoshiKochi opened pull request #2089: resetStyleInheritance property of ShadowRoot is obsolete. (master...a) <a href="https://github.com/w3c/web-platform-tests/pull/2089">https://github.com/w3c/web-platform-tests/pull/2089</a></li>
<li id=l-84 class=hide><a href=#l-84>#</a> [10:35] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-85 class=hide><a href=#l-85>#</a> [10:38] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-86><a href=#l-86>#</a> [10:38] -gitbot:#testing- [web-platform-tests] TakayoshiKochi opened pull request #2090: Remote test for resetStyleInheritance property (master...b) <a href="https://github.com/w3c/web-platform-tests/pull/2090">https://github.com/w3c/web-platform-tests/pull/2090</a></li>
<li id=l-87 class=hide><a href=#l-87>#</a> [10:38] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-88 class=hide _a="heycam|away"><a href=#l-88>#</a> [10:44] * heycam|away is now known as heycam</li>
<li id=l-89 class=hide><a href=#l-89>#</a> [10:49] * Quits: MK (kawamori@public.cloak) (Ping timeout: 180 seconds)</li>
<li id=l-90 class=hide _a="heycam"><a href=#l-90>#</a> [10:59] * heycam is now known as heycam|away</li>
<li id=l-91 class=hide><a href=#l-91>#</a> [11:05] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-92><a href=#l-92>#</a> [11:05] -gitbot:#testing- [web-platform-tests] koustuvsinha opened pull request #2091: Moved animation timing tests (master...move_anim) <a href="https://github.com/w3c/web-platform-tests/pull/2091">https://github.com/w3c/web-platform-tests/pull/2091</a></li>
<li id=l-93 class=hide><a href=#l-93>#</a> [11:05] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-94 class=hide><a href=#l-94>#</a> [11:10] * Joins: MK (kawamori@public.cloak)</li>
<li id=l-95 class=hide><a href=#l-95>#</a> [11:18] * Joins: MK_ (kawamori@public.cloak)</li>
<li id=l-96 class=hide><a href=#l-96>#</a> [11:21] * Quits: MK (kawamori@public.cloak) (Ping timeout: 180 seconds)</li>
<li id=l-97 class=hide><a href=#l-97>#</a> [11:39] * Quits: MK_ (kawamori@public.cloak) (Ping timeout: 180 seconds)</li>
<li id=l-98 class=hide><a href=#l-98>#</a> [11:58] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-99><a href=#l-99>#</a> [11:58] -gitbot:#testing- [web-platform-tests] sideshowbarker closed pull request #2088: Replaced some simple assert_object_equals cases. (master...sideshowbarker/assert_object_equals) <a href="https://github.com/w3c/web-platform-tests/pull/2088">https://github.com/w3c/web-platform-tests/pull/2088</a></li>
<li id=l-100 class=hide><a href=#l-100>#</a> [11:58] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-101 class=hide><a href=#l-101>#</a> [11:58] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-102><a href=#l-102>#</a> [11:58] -gitbot:#testing- [web-platform-tests] sideshowbarker deleted sideshowbarker/assert_object_equals at 1cc2533: <a href="https://github.com/w3c/web-platform-tests/commit/1cc2533">https://github.com/w3c/web-platform-tests/commit/1cc2533</a></li>
<li id=l-103 class=hide><a href=#l-103>#</a> [11:58] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-104 class=hide><a href=#l-104>#</a> [12:02] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-105><a href=#l-105>#</a> [12:02] -gitbot:#testing- [web-platform-tests] Ms2ger closed pull request #2091: Moved animation timing tests (master...move_anim) <a href="https://github.com/w3c/web-platform-tests/pull/2091">https://github.com/w3c/web-platform-tests/pull/2091</a></li>
<li id=l-106 class=hide><a href=#l-106>#</a> [12:02] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-107 class=hide><a href=#l-107>#</a> [12:03] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-108><a href=#l-108>#</a> [12:03] -gitbot:#testing- [web-platform-tests] sideshowbarker created sideshowbarker/assert_object_equals (+1 new commit): <a href="https://github.com/w3c/web-platform-tests/commit/0823b532c7cb">https://github.com/w3c/web-platform-tests/commit/0823b532c7cb</a></li>
<li id=l-109><a href=#l-109>#</a> [12:03] -gitbot:#testing- web-platform-tests/sideshowbarker/assert_object_equals 0823b53 Michael[tm] Smith: Replaced some simple assert_object_equals cases....</li>
<li id=l-110 class=hide><a href=#l-110>#</a> [12:03] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-111 class=hide><a href=#l-111>#</a> [12:04] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-112><a href=#l-112>#</a> [12:04] -gitbot:#testing- [web-platform-tests] sideshowbarker opened pull request #2092: Replaced some simple assert_object_equals cases. (master...sideshowbarker/assert_object_equals) <a href="https://github.com/w3c/web-platform-tests/pull/2092">https://github.com/w3c/web-platform-tests/pull/2092</a></li>
<li id=l-113 class=hide><a href=#l-113>#</a> [12:04] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-114 class=hide><a href=#l-114>#</a> [12:14] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-115><a href=#l-115>#</a> [12:14] -gitbot:#testing- [web-platform-tests] sideshowbarker closed pull request #2092: Replaced some simple assert_object_equals cases. (master...sideshowbarker/assert_object_equals) <a href="https://github.com/w3c/web-platform-tests/pull/2092">https://github.com/w3c/web-platform-tests/pull/2092</a></li>
<li id=l-116 class=hide><a href=#l-116>#</a> [12:14] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-117 class=hide><a href=#l-117>#</a> [12:15] * Quits: Lachy (~Lachy@public.cloak) (&quot;My MacBook Pro has gone to sleep. ZZZzzz…&quot;)</li>
<li id=l-118 class=hide><a href=#l-118>#</a> [12:29] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-119><a href=#l-119>#</a> [12:29] -gitbot:#testing- [web-platform-tests] TakayoshiKochi opened pull request #2093: Fix createTestMediaPlayer (master...events) <a href="https://github.com/w3c/web-platform-tests/pull/2093">https://github.com/w3c/web-platform-tests/pull/2093</a></li>
<li id=l-120 class=hide><a href=#l-120>#</a> [12:29] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-121 class=hide><a href=#l-121>#</a> [12:49] * Joins: Lachy (~Lachy@public.cloak)</li>
<li id=l-122 class=hide><a href=#l-122>#</a> [13:45] * Joins: jyasskin (~textual@public.cloak)</li>
<li id=l-123 class=hide><a href=#l-123>#</a> [13:46] * Quits: jyasskin (~textual@public.cloak) (&quot;My computer has gone to sleep. ZZZzzz…&quot;)</li>
<li id=l-124 class=hide><a href=#l-124>#</a> [13:57] * Quits: gamekathu (~gamekathu@public.cloak) (&quot;Page closed&quot;)</li>
<li id=l-125 class=hide><a href=#l-125>#</a> [13:58] * Joins: MK (kawamori@public.cloak)</li>
<li id=l-126 class=hide><a href=#l-126>#</a> [14:13] * Quits: MK (kawamori@public.cloak) (Ping timeout: 180 seconds)</li>
<li id=l-127 class=hide><a href=#l-127>#</a> [15:30] * Joins: MK (kawamori@public.cloak)</li>
<li id=l-128 class=hide><a href=#l-128>#</a> [15:45] * Quits: MK (kawamori@public.cloak) (Ping timeout: 180 seconds)</li>
<li id=l-129 class=hide><a href=#l-129>#</a> [16:13] * Joins: MK (kawamori@public.cloak)</li>
<li id=l-130 class=hide><a href=#l-130>#</a> [16:17] * Quits: Ms2ger (~Ms2ger@public.cloak) (Ping timeout: 180 seconds)</li>
<li id=l-131 class=hide><a href=#l-131>#</a> [16:38] * Joins: Ms2ger (~Ms2ger@public.cloak)</li>
<li id=l-132 class=hide><a href=#l-132>#</a> [17:05] * Joins: bhill2 (~bhill2@public.cloak)</li>
<li id=l-133 class=hide><a href=#l-133>#</a> [17:24] * Joins: jyasskin (~textual@public.cloak)</li>
<li id=l-134 class=hide><a href=#l-134>#</a> [17:47] * Quits: dom (dom@public.cloak) (&quot;&quot;)</li>
<li id=l-135 class=hide><a href=#l-135>#</a> [17:55] * Quits: MK (kawamori@public.cloak) (Ping timeout: 180 seconds)</li>
<li id=l-136 class=hide><a href=#l-136>#</a> [18:09] * Quits: Lachy (~Lachy@public.cloak) (&quot;My MacBook Pro has gone to sleep. ZZZzzz…&quot;)</li>
<li id=l-137 class=hide><a href=#l-137>#</a> [18:16] * Quits: Ms2ger (~Ms2ger@public.cloak) (Ping timeout: 180 seconds)</li>
<li id=l-138 class=hide><a href=#l-138>#</a> [18:23] * Quits: bhill2 (~bhill2@public.cloak) (Client closed connection)</li>
<li id=l-139 class=hide><a href=#l-139>#</a> [18:28] * Joins: bhill2 (~bhill2@public.cloak)</li>
<li id=l-140 class=hide><a href=#l-140>#</a> [18:35] * Joins: Lachy (~Lachy@public.cloak)</li>
<li id=l-141 class=hide><a href=#l-141>#</a> [18:48] * Quits: Lachy (~Lachy@public.cloak) (&quot;My MacBook Pro has gone to sleep. ZZZzzz…&quot;)</li>
<li id=l-142 class=hide><a href=#l-142>#</a> [20:40] * Quits: jyasskin (~textual@public.cloak) (&quot;My computer has gone to sleep. ZZZzzz…&quot;)</li>
<li id=l-143 class=hide><a href=#l-143>#</a> [20:51] * Joins: Lachy (~Lachy@public.cloak)</li>
<li id=l-144 class=hide><a href=#l-144>#</a> [21:09] * Joins: cvrebert (~cvrebert@public.cloak)</li>
<li id=l-145 _a="cvrebert"><a href=#l-145>#</a> [21:10] &lt;cvrebert&gt; Who owns W3C Sync Bot (<a href="https://github.com/syncbot">https://github.com/syncbot</a> ) or where do I file a bug against it?</li>
<li id=l-146 class=hide><a href=#l-146>#</a> [21:11] * Quits: Lachy (~Lachy@public.cloak) (&quot;My MacBook Pro has gone to sleep. ZZZzzz…&quot;)</li>
<li id=l-147 class=hide><a href=#l-147>#</a> [21:38] * Joins: Lachy (~Lachy@public.cloak)</li>
<li id=l-148 class=hide><a href=#l-148>#</a> [21:50] * Quits: Lachy (~Lachy@public.cloak) (&quot;My MacBook Pro has gone to sleep. ZZZzzz…&quot;)</li>
<li id=l-149 class=hide><a href=#l-149>#</a> [22:00] * Quits: bhill2 (~bhill2@public.cloak) (Client closed connection)</li>
<li id=l-150 class=hide><a href=#l-150>#</a> [22:03] * Joins: bhill2 (~bhill2@public.cloak)</li>
<li id=l-151 class=hide><a href=#l-151>#</a> [22:04] * Quits: bhill2 (~bhill2@public.cloak) (Client closed connection)</li>
<li id=l-152 _a="jgraham" _r="plinss"><a href=#l-152>#</a> [22:06] &lt;jgraham&gt; plinss: ^</li>
<li id=l-153 _a="plinss" _r="cvrebert"><a href=#l-153>#</a> [22:07] &lt;plinss&gt; cvrebert: yes, that’s mine, what’s the issue?</li>
<li id=l-154 _a="cvrebert" _r="plinss"><a href=#l-154>#</a> [22:09] &lt;cvrebert&gt; plinss: Gerard's feedback in <a href="https://lists.w3.org/Archives/Public/public-css-testsuite/2015Aug/0003.html">https://lists.w3.org/Archives/Public/public-css-testsuite/2015Aug/0003.html</a> (and the lack of corresponding complaints from the bot on <a href="https://github.com/w3c/csswg-test/pull/813">https://github.com/w3c/csswg-test/pull/813</a> ) suggests there's a couple things the bot ought to check that it currently isn't</li>
<li id=l-155 _a="cvrebert" _r="plinss"><a href=#l-155>#</a> [22:10] &lt;cvrebert&gt; plinss: (1) <a href="https://drafts.csswg.org">https://drafts.csswg.org</a> shouldn't be used in rel=&quot;help&quot; links</li>
<li id=l-156 _a="plinss" _r="actually"><a href=#l-156>#</a> [22:11] &lt;plinss&gt; actually, that’s not an issue, draft links are fine</li>
<li id=l-157 _a="cvrebert" _r="plinss"><a href=#l-157>#</a> [22:12] &lt;cvrebert&gt; plinss: (2) testcase filenames should be unique (irrespective of subdirectory apparently?); see <a href="https://github.com/w3c/web-platform-tests/issues/2095">https://github.com/w3c/web-platform-tests/issues/2095</a></li>
<li id=l-158 _a="plinss" _r="yeah"><a href=#l-158>#</a> [22:13] &lt;plinss&gt; yeah, they should be (test stutes can get built from multiple diectoried and the dir structure isn’t preserved)</li>
<li id=l-159 _a="plinss"><a href=#l-159>#</a> [22:14] &lt;plinss&gt; I can add a check for that</li>
<li id=l-160 _a="cvrebert"><a href=#l-160>#</a> [22:14] &lt;cvrebert&gt; great</li>
<li id=l-161 _a="cvrebert" _r="plinss"><a href=#l-161>#</a> [22:15] &lt;cvrebert&gt; plinss: also, it would be nice if <a href="https://github.com/syncbot">https://github.com/syncbot</a> 's profile had its URL field set to something useful, so folks don't need to chase down who to contact about these sorts of things</li>
<li id=l-162 _a="plinss"><a href=#l-162>#</a> [22:16] &lt;plinss&gt; good point, will fix that too</li>
<li id=l-163 class=hide><a href=#l-163>#</a> [22:18] * Joins: bhill2 (~bhill2@public.cloak)</li>
<li id=l-164 class=hide><a href=#l-164>#</a> [22:21] * Joins: Lachy (~Lachy@public.cloak)</li>
<li id=l-165 class=hide><a href=#l-165>#</a> [22:48] * Quits: cvrebert (~cvrebert@public.cloak) (cvrebert)</li>
<li id=l-166 class=hide><a href=#l-166>#</a> [22:49] * Quits: bhill2 (~bhill2@public.cloak) (Client closed connection)</li>
<li id=l-167 class=hide><a href=#l-167>#</a> [22:50] * Joins: bhill2_ (~bhill2@public.cloak)</li>
<li id=l-168 class=hide><a href=#l-168>#</a> [23:54] * Quits: Lachy (~Lachy@public.cloak) (&quot;My MacBook Pro has gone to sleep. ZZZzzz…&quot;)</li>
<li id=l-169 class=hide><a href=#l-169>#</a> Session Close: Tue Aug 18 00:00:00 2015</li>
</ol>
<p id=end><a href=20150816>Previous day</a>, <a href=20150818>Next day</a></p>
<p>Think <a href="/irc-logs/">these logs</a> are useful? Then <a href="https://www.paypal.com/cgi-bin/webscr?cmd=_donations&amp;business=PMXKLAEH53JSS&amp;lc=GB&amp;item_name=Krijn%20Hoetmer&amp;item_number=IRC%20Logs&amp;currency_code=EUR&amp;bn=PP%2dDonationsBF%3abtn_donate_SM%2egif%3aNonHosted"><strong>please donate</strong></a> to show your gratitude (and keep them up, of course). <a href="https://krijnhoetmer.nl/">Thanks! — Krijn</a></p>
<script src=/irc-logs/s2.js></script>
